Apply some fat commas to commands of TAP tests
This is similar to 19c6e92b13b2, in order to keep the style used in the
scripts consistent for the option names and values used in commands.
The places updated in this commit have been added recently in
71ea0d679543.

These changes are cosmetic; there is no need for a backpatch.